+++ frontend/src/store/modules/index.js
@@ -5,3 +5,4 @@ export { useShowNotify } from "./useShowNotify";
 export { useNotifyList } from "./useNotifyList";
 export { useSearchProduct } from "./useSearchProduct";
 export { useUserStore } from "./useUserStore";
+export { useMypageTab } from "./useMypageTab";
