+++ frontend/src/store/modules/useMypageTab.js
@@ -1,0 +1,14 @@
+import { defineStore } from "pinia";
+
+export const useMypageTab = defineStore("mypageTab", {
+  state: () => {
+    return {
+      tabIndex: 0,
+    };
+  },
+  actions: {
+    setTabIndex(index) {
+      this.tabIndex = index;
+    },
+  },
+});
