+++ frontend/src/views/mypage/index.js
@@ -1,2 +1,5 @@
-export { default as MypageView } from "./MypageView.vue";
-export { default as EditProfileView } from "./EditProfileView.vue";
+export { default as MypageView } from "./MyPageView.vue";
+export { default as MypageHomeView } from "./MypageHomeView.vue";
+export { default as MypageEditProfileView } from "./MypageEditProfileView.vue";
+export { default as MypagePurchaseListView } from "./MyPagePurchaseListView.vue";
+export { default as MypageSellListView } from "./MyPageSellListView.vue";
